[PATCH 1629/3839] Fix to FormRenderer
turns out that FormRenderer does not support the input of all data types
in particular nested generics used to cause an exception.
This change makes so that nested generics are tolerated, i.e. they are
not rendered in the form but they don't prevent the rest of the form from
being rendered.

Also, added a test to show the new behavior.

RB_ID=85009

(sapling split of 3a20a711e04612dced8de924e21372b4e9537b6b)