[PATCH 3345/3839] Strainer Implementation.
This patch is the Strainer library, a little language and engine for
filtering messages on rosette. A filter might look like the following:
((a.b > 0) AND (a.c matches "a(b|c)d")). Each atom is composed of a
field name, operator and argument. Expressions are composed of ANDing
and ORing atoms/expressions together. The field name is dotted to
represented nested access in a JSON or Thrift object.

This patch has the following components:
- Strainer parser and lexer rules specified in ANTLR 4.
- Execution Engine for taking in a JSON object and applying a strainer filter to it.
- Tests for the parser, lexer, and execution engine.

There are three sets of tests:
- 'EventSuite', tests for mapping a strainer field name to a JSON object.
- 'ParsingSuite', tests to check the parser and lexer.
- 'CompilingSuite', tests that apply a filter to a JSON object and
  ensure the filter was apply correctly.

RB_ID=122455

(sapling split of 8e8c866139cfba1c51a05ff8e15cb799fe70826d)