[PATCH 4771/7508] zookeeper: Add a queue to process WatchedEvent in order
Motivation:
Finagle's StabilizingAddr (formerly known as StabilizingGroup) has a notion of health of a cluster. This health is determine by the "zookeeper state" of the events it receives. If the ordering of those events is changed, then it could received 'SyncConnected' before 'Expired' and then consider the cluster as unhealthy.

Modification:
Introduced a "Processor" thread, which is responsible for processing WatchedEvent and a LinkedBlockedQueue.
The "SessionWatcher" is now sending events in the queue and those events are processed by the "Processor", if processing one of those events generate a new event, this event is enqueue AT THE END (and not processed immediately).

Result:
Every events are received in order by the Watcher.
StabilizingAddr perceived correctly the ZK cluster as (un)healthy.

RB_ID=288321

(sapling split of ee51df6d2ad70446a8b669b331204d04f78ae396)